+++ src/classes/game-settings.ts
@@ -5,7 +5,8 @@ import {
     GeneralSettings,
     LeapYearConfig,
     MonthConfig,
-    MoonConfiguration, NoteCategory,
+    MoonConfiguration,
+    NoteCategory,
     NoteConfig,
     SeasonConfiguration,
     SimpleCalendarSocket,
@@ -353,8 +354,9 @@ export class GameSettings {
     /**
      * Saves the current date to the world settings
      * @param {Year} year The year that has the current date
+     * @param {boolean} emitHook If to emit the hook or not
      */
-    static async SaveCurrentDate(year: Year): Promise<boolean>{
+    static async SaveCurrentDate(year: Year, emitHook: boolean = true): Promise<boolean>{
         if(GameSettings.IsGm()){
             Logger.debug(`Saving current date.`);
             const currentMonth = year.getMonth();
@@ -370,7 +372,9 @@ export class GameSettings {
                     };
                     if(currentDate.year !== newDate.year || currentDate.month !== newDate.month || currentDate.day !== newDate.day || currentDate.seconds !== newDate.seconds){
                         await (<Game>game).settings.set(ModuleName, SettingNames.CurrentDate, newDate);
-                        Hook.emit(SimpleCalendarHooks.DateTimeChange);
+                        if(emitHook){
+                            Hook.emit(SimpleCalendarHooks.DateTimeChange);
+                        }
                         await GameSockets.emit(<SimpleCalendarSocket.Data>{ type: SocketTypes.noteReminders, data: { justTimeChange: currentDate.seconds !== newDate.seconds && currentDate.day === newDate.day }});
                         if(SimpleCalendar.instance){
                             SimpleCalendar.instance.checkNoteReminders(currentDate.seconds !== newDate.seconds && currentDate.day === newDate.day);